+++ src/api/weatherAPI.ts
@@ -1,0 +1,6 @@
+import { instance } from "./api";
+
+export const getWeather = async () => {
+  const response = await instance.get(`/location`);
+  return response.data;
+};
