+++ src/pages/Map/Maps.tsx
@@ -2,17 +2,29 @@ import { useState, useEffect } from "react";
 import { Map } from "react-kakao-maps-sdk";
 import { getWeather } from "../../api/weatherAPI";
 import { MARKER_DATA } from "../../data/MARKER_DATA";
-import WeatherOverlay from "./components/WeatherOverlay/WeatherOverlay";
+import WeatherOverlay, {
+  Root,
+} from "./components/WeatherOverlay/WeatherOverlay";
 import Buttons from "./components/Buttons/Buttons";
 import Checkbox from "./components/Checkbox/Checkbox";
 import axios from "axios";
-import { getMicrodustsLevel } from "../../api/microdustsAPI";
-import Microdusts from "./components/MicrodustsOverlay/MicrodustsOverlay";
+import { useQuery } from "@tanstack/react-query";
+import Loading from "../../components/Loading/Loading";
+import { MapsDataType } from "./Types/MapsDataTypes";
 
 const Maps = () => {
   const [level, setLevel] = useState(13);
   const [dataSwitch, setDataSwitch] = useState(true);
+  // const [temperature, setTemperature] = useState();
+  const [icon, setIcon] = useState<string>();
 
+  const {
+    status,
+    data: { data },
+  } = useQuery(["weatherData"], () => getWeather());
+  console.log(data);
+
+  if (status === "loading") return <Loading />;
   // useEffect(() => {
   //   getWeather(35.17989493738095, 129.07481938748694).then((response) =>
   //   );
@@ -26,8 +38,34 @@ const Maps = () => {
   };
 
   // useEffect(() => {
-  //   getMicrodustsLevel(4).then((response) => console.log(response.data));
-  // });
+  //   axios
+  //     .get(
+  //       `https://api.openweathermap.org/data/2.5/weather?lat=${lat}&lon=${lng}&units=metric&appid=${apiKey}`
+  //     )
+  //     .then((res) => {
+  //       let weatherIcons;
+  //       switch (res.data.weather[0].main) {
+  //         case "Clear":
+  //           weatherIcons = sun;
+  //           break;
+  //         case "Clouds":
+  //           weatherIcons = clouds;
+  //           break;
+  //         case "Rain":
+  //           weatherIcons = rain;
+  //           break;
+  //         case "Snow":
+  //           weatherIcons = snow;
+  //           break;
+  //         default:
+  //           weatherIcons = sun;
+  //       }
+  //       setTemperature(res.data);
+  //       setIcon(weatherIcons);
+  //     });
+
+  //   // api.getWeather(lat, lng).then((res) => console.log(res.data));
+  // }, []);
 
   return (
     <Map
@@ -35,13 +73,9 @@ const Maps = () => {
       style={{ width: "100%", height: "1000px", position: "relative" }}
       level={level}
     >
-      {dataSwitch ? (
-        MARKER_DATA.map((city) => (
-          <WeatherOverlay data={city} key={city.title} />
-        ))
-      ) : (
-        <Microdusts />
-      )}
+      {data.map((weather: MapsDataType) => (
+        <WeatherOverlay data={weather} />
+      ))}
 
       <Buttons
         weatherSwitch={weatherSwitch}
